move state to context
