+++ src/uTechpia/context/uTechContext.jsx
@@ -1,10 +1,41 @@
-import React, { createContext, useState } from 'react';
+import React, { createContext, useState, useEffect } from 'react';
+import dataJson from '../../utils/data';
+import { getDirectRuleOfThree } from '../../utils/usefulFunctions';
 
 export const ContextUTech = createContext();
 
 export const ProviderUTech = props => {
-  const [prueba, setPrueba] = useState('');
+  const [data, setData] = useState(dataJson);
+  const [screenSize, setScreenSize] = useState({
+    width: 1920,
+    height: 1080,
+  });
+
+  useEffect(() => {
+    const screenHeight = window.innerHeight;
+    const screenWidth = window.innerWidth;
+    console.log({ screenHeight, screenWidth });
+    setScreenSize({
+      width: screenHeight,
+      height: screenWidth,
+    });
+
+    const responsiveData = data.map(dot => {
+      return {
+        ...dot,
+        radius: getDirectRuleOfThree(1920, screenWidth, dot.radius),
+        position: {
+          x: getDirectRuleOfThree(1920, screenWidth, dot.position.x),
+          y: getDirectRuleOfThree(1080, screenHeight, dot.position.y),
+        },
+      };
+    });
+    setData(responsiveData);
+  }, []);
+
   return (
-    <ContextUTech.Provider value={{ prueba, setPrueba }}>{props.children}</ContextUTech.Provider>
+    <ContextUTech.Provider value={{ data, setData, screenSize, setScreenSize }}>
+      {props.children}
+    </ContextUTech.Provider>
   );
 };
